refactor tests: unit to integration level (#198)
