+++ cargo-smart-release/tests/changelog/mod.rs
@@ -1,0 +1,13 @@
+// mod parse {
+//     use cargo_smart_release::ChangeLog;
+//
+//     #[test]
+//     #[ignore]
+//     fn all_unknown_in_section() {
+//         // ChangeLog::from_markdown(include_str!())
+//     }
+// }
+
+mod write_and_parse;
+
+mod merge;
