+++ cargo-smart-release/tests/integration.rs
@@ -1,0 +1,3 @@
+type Result = std::result::Result<(), Box<dyn std::error::Error>>;
+
+mod changelog;
